Adds a new view for stats
Implements the get_stats method in the Stats
Controller. It checks if the token exists in the hash
and returns it if it does, else computes the value, adds it to hash
and returns it.

Makes an ajax call(POST) to tools/stats.json that invokes the get_stats
method in the controller.

Renames a migration file.